+++ Src/Animation/AnimBrownian.cs
@@ -5,7 +5,7 @@ namespace WebGal.Animations;
 class AnimationBrownian : IAnimation
 {
 	readonly Random rand = new();
-	public FVector GetOffset(float timeOff, long timeObs)
+	public FVector GetOffset(double timeOff, long timeObs)
 	{
 		var (x, y) = (rand.NextDouble(), rand.NextDouble());
 		return new((float)x, (float)y);
