+++ Src/Animation/IAnimation.cs
@@ -4,5 +4,11 @@ namespace WebGal.Animations;
 
 public interface IAnimation
 {
-	public FVector GetOffset(float timeOff, long timeObs);
+	/// <summary>
+	/// 
+	/// </summary>
+	/// <param name="timeOff">时间偏移量，秒为单位</param>
+	/// <param name="timeObs">时间绝对值，ms为单位</param>
+	/// <returns></returns>
+	public FVector GetOffset(double timeOff, long timeObs);
 }
